Workers module, stop passing them around
