+++ src/chain.js
@@ -1,5 +1,5 @@
 var JobPackager = require('./job_packager');
-var ResultCollector = require('./result_collector');
+var workers = require('./workers');
 var merge_typed_arrays = require('./typed_array_merger');
 var operation_names = require('./operation_names');
 var operation_packager = require('./operation_packager');
@@ -32,11 +32,10 @@ var finisher = {
   }
 };
 
-var Chain = function (source, parts, workers, operation, globalContext, chainContext, previousOperations) {
+var Chain = function (source, parts, operation, globalContext, chainContext, previousOperations) {
   this.packager = new JobPackager(parts, source);
   this.source = source;
   this.parts = parts;
-  this.workers = workers;
   this.operation = operation; //todo: (mati) por ahora lo dejo para finalizar el reduce correctamente
   this.globalContext = globalContext;
   this.chainContext = chainContext;
@@ -53,14 +52,14 @@ Chain.prototype.map = function (mapper, localContext) {
   this.__verifyPreviousOperation();
   var extendChainContext = contextUtils.extendChainContext(localContext, this.chainContext);
   var operation = operation_packager(operation_names.MAP, mapper);
-  return new Chain(this.source, this.parts, this.workers, operation, this.globalContext, extendChainContext, this.operations);
+  return new Chain(this.source, this.parts, operation, this.globalContext, extendChainContext, this.operations);
 };
 
 Chain.prototype.filter = function (predicate, localContext) {
   this.__verifyPreviousOperation();
   var extendChainContext = contextUtils.extendChainContext(localContext, this.chainContext);
   var operation = operation_packager(operation_names.FILTER, predicate);
-  return new Chain(this.source, this.parts, this.workers, operation, this.globalContext, extendChainContext, this.operations);
+  return new Chain(this.source, this.parts, operation, this.globalContext, extendChainContext, this.operations);
 };
 
 Chain.prototype.reduce = function (reducer, seed, identityReducer, identity, localContext) {
@@ -72,41 +71,21 @@ Chain.prototype.reduce = function (reducer, seed, identityReducer, identity, loc
   this.__verifyPreviousOperation();
   var extendChainContext = contextUtils.extendChainContext(localContext, this.chainContext);
   var operation = operation_packager(operation_names.REDUCE, reducer, seed, identity, identityReducer);
-  return new Chain(this.source, this.parts, this.workers, operation, this.globalContext, extendChainContext, this.operations);
+  return new Chain(this.source, this.parts, operation, this.globalContext, extendChainContext, this.operations);
 };
 
 Chain.prototype.seq = function (done) {
   var self = this;
-  var workers = this.workers;
   var TypedArrayConstructor = this.source.constructor;
   var packs = this.packager.generatePackages(this.operations, this.chainContext);
-  var collector = new ResultCollector(this.parts, function(err, results){
+
+  workers.sendPacks(packs, function(err, results){
     if (err) { return done(err); }
     var partial_results = results.map(function(result){
       return new TypedArrayConstructor(result.value).subarray(0, result.newLength);
     });
     var m = merge_typed_arrays(partial_results);
     return finisher[self.operation.name](self, m, done);
-  });
-
-  packs.forEach(function(pack, index){
-    var onMessageHandler = function (event){
-      event.target.removeEventListener('error', onErrorHandler);
-      event.target.removeEventListener('message', onMessageHandler);
-      return collector.onPart(event.data);
-    };
-
-    var onErrorHandler = function (event){
-      event.preventDefault();
-      event.target.removeEventListener('error', onErrorHandler);
-      event.target.removeEventListener('message', onMessageHandler);
-      return collector.onError(event.message);
-    };
-
-    workers[index].addEventListener('error', onErrorHandler);
-    workers[index].addEventListener('message', onMessageHandler);
-
-    workers[index].postMessage(pack, [ pack.buffer ]);
   });
 };
