+++ src/wrapped_typed_array.js
@@ -4,10 +4,9 @@ var operation_packager = require('./operation_packager');
 var contextUtils = require('./chain_context');
 var utils = require('./utils');
 
-var WrappedTypedArray = function (source, parts, workers, globalContext) {
+var WrappedTypedArray = function (source, parts, globalContext) {
   this.source = source;
   this.parts = parts;
-  this.workers = workers;
   this.globalContext = globalContext;
 };
 
@@ -31,7 +30,7 @@ WrappedTypedArray.prototype.reduce = function(reducer, seed, identityReducer, id
 WrappedTypedArray.prototype.__operation = function(name, code, localContext, seed, identity, identityCode) {
   var operation = operation_packager(name, code, seed, identity, identityCode);
   var chainContext = contextUtils.extendChainContext(localContext);
-  return new Chain(this.source, this.parts, this.workers, operation, this.globalContext, chainContext);
+  return new Chain(this.source, this.parts, operation, this.globalContext, chainContext);
 };
 
 module.exports = WrappedTypedArray;
